server/requirements.txt
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
build==1.0.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.2
django-allauth==0.61.1
django-appconf==1.0.6
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-filter==23.5
django-imagekit==5.0.0
django-js-asset==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
fastjsonschema==2.19.1
fqdn==1.5.1
idna==3.6
isoduration==20.11.0
Jinja2==3.1.3
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
Markdown==3.5.2
MarkupSafe==2.1.5
mistune==3.0.2
mysqlclient==2.2.4
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
oauthlib==3.2.2
overrides==7.7.0
packaging==23.2
pandocfilters==1.5.1
pilkit==3.0
pillow==10.2.0
pip-autoremove==0.10.0
pip-tools==7.4.0
platformdirs==4.2.0
prometheus_client==0.20.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
terminado==0.18.1
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.2
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
